move tests to test folder and remove sdk dep
